# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: >-
    Healers Inc marketplace API. Every timestamp crossing this boundary is a UTC
    instant (RFC 3339 with a `Z` suffix); wall-clock rendering is always derived
    from an IANA time zone supplied alongside it. Errors are returned as
    `{ "error": string, "message": string }` with a 4xx/5xx status.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: Appointments for the signed-in account in its active role
      parameters:
        - name: scope
          in: query
          schema:
            $ref: "#/components/schemas/AppointmentScope"
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createBooking
      tags: [appointments]
      summary: Book a session
      description: >-
        Availability is re-checked inside the writing transaction and the
        database rejects any overlap, so a slot cannot be sold twice. Repeating
        the same `idempotencyKey` returns the original appointment instead of
        creating a second one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: A single appointment
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}/actions:
    post:
      operationId: actOnAppointment
      tags: [appointments]
      summary: Accept, decline, cancel, reschedule or complete an appointment
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentActionRequest"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}/checkout:
    post:
      operationId: createAppointmentCheckout
      tags: [appointments]
      summary: Create or reuse hosted checkout for an accepted appointment
      description: >-
        Checkout is available only after practitioner acceptance. The server
        supplies allowlisted native or web success and cancellation
        destinations based on the requested return target.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCheckoutInput"
      responses:
        "200":
          description: Hosted checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
  /appointments/{appointmentId}/payment:
    get:
      operationId: getAppointmentPayment
      tags: [appointments]
      summary: Payment state for an appointment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      responses:
        "200":
          description: Appointment payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSummary"
  /appointments/{appointmentId}/refunds:
    post:
      operationId: createAppointmentRefund
      tags: [appointments]
      summary: Request a full appointment refund
      description: >-
        Cancellations before the practitioner's notice window receive a full
        refund. Late cancellations are not refunded automatically.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRefundInput"
      responses:
        "201":
          description: Refund requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefundSummary"
  /appointments/{appointmentId}/ics:
    get:
      operationId: getAppointmentCalendarFile
      tags: [appointments]
      summary: Calendar file for the appointment
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      responses:
        "200":
          description: iCalendar payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarFile"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    AppointmentId:
      name: appointmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    SessionFormat:
      type: string
      enum: [online, phone, in_person]
    PortalRole:
      type: string
      enum: [client, practitioner]
...
    AppointmentStatus:
      type: string
      enum: [pending, confirmed, completed, cancelled, declined]
...
    RefundStatus:
      type: string
      enum: [pending, succeeded, failed, cancelled]
    AppointmentScope:
      type: string
      enum: [upcoming, pending, past, all]
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        serviceName:
          type: string
        serviceDurationMinutes:
          type: integer
        servicePriceCents:
          type: integer
        serviceCurrency:
          type: string
        serviceFormat:
          $ref: "#/components/schemas/SessionFormat"
        clientTimezone:
          type: string
        practitionerTimezone:
          type: string
        viewerTimezone:
          type: string
        viewerLabel:
          type: string
        counterpartLabel:
          type: string
        clientNotes:
          type: ["string", "null"]
        cancellationReason:
          type: ["string", "null"]
        cancelledLate:
          type: boolean
        client:
          $ref: "#/components/schemas/PartyRef"
        practitioner:
          $ref: "#/components/schemas/PartyRef"
        viewerRole:
          $ref: "#/components/schemas/PortalRole"
        canReview:
          type: boolean
        hasReview:
          type: boolean
        payment:
          oneOf:
            - $ref: "#/components/schemas/PaymentSummary"
            - type: "null"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - startsAt
        - endsAt
        - serviceName
        - serviceDurationMinutes
        - servicePriceCents
        - serviceCurrency
        - serviceFormat
        - clientTimezone
        - practitionerTimezone
        - viewerTimezone
        - viewerLabel
        - counterpartLabel
        - cancelledLate
        - client
        - practitioner
        - viewerRole
        - canReview
        - hasReview
        - payment
        - createdAt
    PaymentSummary:
      type: object
      properties:
        id:
          type: ["string", "null"]
          format: uuid
        appointmentId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/PractitionerPaymentStatus"
        message:
          type: ["string", "null"]
        grossAmountCents:
          type: integer
        platformFeeCents:
          type: integer
        platformFeeBasisPoints:
          type: integer
          description: Platform fee policy captured when checkout was created.
        practitionerNetCents:
          type: integer
        currency:
          type: string
        paidAt:
          type: ["string", "null"]
          format: date-time
        refundedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - appointmentId
        - status
        - message
        - grossAmountCents
        - platformFeeCents
        - platformFeeBasisPoints
        - practitionerNetCents
        - currency
        - paidAt
        - refundedAt
        - createdAt
...
    CheckoutSessionResponse:
      type: object
      properties:
        payment:
          $ref: "#/components/schemas/PaymentSummary"
        checkoutUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
      required:
        - payment
        - checkoutUrl
        - expiresAt
    RefundSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        paymentId:
          type: string
          format: uuid
        amountCents:
          type: integer
        currency:
          type: string
        status:
          $ref: "#/components/schemas/RefundStatus"
        reason:
          type: ["string", "null"]
        refundedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - paymentId
        - amountCents
        - currency
        - status
        - reason
        - refundedAt
        - createdAt
...
    PaymentCheckoutInput:
      type: object
      properties:
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
        returnTarget:
          type: string
          description: >-
            Where hosted Checkout should send the payer back to. `native` uses
            the app's deep-link handlers, `web` the Expo web build, and `site`
            the Healers Inc website. The server composes every destination
            itself from allowlisted configuration; callers never supply a URL.
          enum:
            - native
            - web
            - site
      required:
        - idempotencyKey
        - returnTarget
    PaymentRefundInput:
      type: object
      properties:
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
        reason:
          type: ["string", "null"]
      required:
        - idempotencyKey
    BookingRequest:
      type: object
      properties:
        practitionerId:
          type: string
        serviceId:
          type: string
        startsAt:
          type: string
          format: date-time
        clientTimezone:
          type: string
        notes:
          type: ["string", "null"]
        idempotencyKey:
          type: string
          minLength: 8
      required:
        - practitionerId
        - serviceId
        - startsAt
        - clientTimezone
        - idempotencyKey
    AppointmentActionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [accept, decline, cancel, complete, reschedule]
        startsAt:
          type: ["string", "null"]
          format: date-time
          description: Required for `reschedule`
        reason:
          type: ["string", "null"]
      required:
        - action
    CalendarFile:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
        content:
          type: string
      required:
        - filename
        - contentType
        - content
...